refactor: move types to a separated file
