+++ modules/kolours/types/index.ts
@@ -1,0 +1,3 @@
+import * as Kolours from "./Kolours";
+
+export { Kolours };
